moved async functions to helper file
